decoder: CB1 fixed playback stuck after seek. [1/1]
PD#SWPL-68587

Problem:
Since the vdec_core_finish_run function is called twice, the status will
eventually be set to connect. During the reset connect process, it will
return if the status is not equal to disconnect, and the playback will
be stuck if there is no decoding instance.

Solution:
The timer function does not need to schedule work after detecting the
decoding state to prevent two work from appearing in one run.

Verify:
T950D4

Change-Id: Iddcaac752c4199202e549696bec19ef9310e1e0c
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>